+++ analytics.js
@@ -1,4 +1,4 @@
-//     Analytics.js 0.4.3
+//     Analytics.js 0.4.4
 
 //     (c) 2013 Segment.io Inc.
 //     Analytics.js may be freely distributed under the MIT license.
@@ -1462,11 +1462,13 @@ analytics.addProvider('Mixpanel', {
         // Alias the traits' keys with dollar signs for Mixpanel's API.
         if (traits) {
             analytics.utils.alias(traits, {
-                'email'    : '$email',
-                'name'     : '$name',
-                'username' : '$username',
-                'lastSeen' : '$lastSeen',
-                'created'  : '$created'
+                'created'   : '$created',
+                'email'     : '$email',
+                'firstName' : '$first_name',
+                'lastName'  : '$last_name',
+                'lastSeen'  : '$last_seen',
+                'name'      : '$name',
+                'username'  : '$username'
             });
         }